refactor(usePolicy): RHICOMPL-1624 move and split to Mutations
