+++ src/Mutations/index.js
@@ -1,1 +1,2 @@
 export * from './graphql/mutations';
+export { default as usePolicy } from './hooks/usePolicy';
